fix type

+++ service/edit_version__service.go
@@ -17,7 +17,7 @@ type EditVersionService struct {
 }
 
 func (v *EditVersionService) EditVersion() response.Response {
-	version, err := mysql.EditVersion(v.Version, v.URL, v.Desc, v.CodeID, v.ID)
+	version, err := mysql.EditVersion(v.Version, v.URL, v.Desc, int64(v.CodeID), v.ID)
 	if err != nil {
 		util.Log().Error("edit version err: %v", err)
 		return response.NewResponse(errcode.CodeDBError, nil, errcode.Text(errcode.CodeDBError))
